requirements.txt
prefect==2.14.21
pendulum<3.0.0
psycopg2-binary==2.9.9
pandas==2.2.1
SQLAlchemy==2.0.29
requests==2.31.0
python-dotenv==1.0.1

